Modified route planner classes a little
The child class constructor was set up to optionally receive input, but
this will never happen. One of the inputs was the data that the route
planner is actually defining. The other optional input was the data that
the route planner uses to operate, and this is defined in the facade.
Both private variables were moved to the parent class because they will
be common to any type of route planner.