Modified two unit tests related to lane splits
The scenario in the tests was modified for clarity. Instead of modifying
the lane end cost of only the straight connection at the split, the lane
end costs for both lanes continuing straight were modified. Now, these
tests more accurately represent a scenario where the vehicle needs to
take the split on the right to exit.