Modified GFF creation for adjacent lanes
1. Previously, GFFs for the adjacent lanes were created starting from
the host's station value in its lane. Since identical station values in
adjacent lanes are not necessarily next to each other, the host's
equivalent station value in the adjacent lanes was found and used to
create the adjacent lane GFFs.
2. The arguments for MapUtils.get_lookahead_frenet_frame_by_cost were
modified. The function now just takes the host's equivalent station in
each lane.
3. A couple unit tests were modified to account for these changes.